+++ src/main/java/com/projectjesusangels/entity/Apoderado.java
@@ -18,7 +18,7 @@ public class Apoderado {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_apoderado")
-    private Long idApoderado;
+    private Integer idApoderado;
 
     @Column(name = "nombre", nullable = false, length = 100)
     private String nombre;
