+++ src/main/java/com/projectjesusangels/entity/Curso.java
@@ -15,7 +15,7 @@ public class Curso {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_curso")
-    private Long idCurso;
+    private Integer idCurso;
 
     @Column(name = "nombre", nullable = false, unique = true, length = 100)
     private String nombre;
