+++ src/main/java/com/projectjesusangels/repository/PagoRepository.java
@@ -9,8 +9,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PagoRepository extends JpaRepository<Pago, Long> {
-    List<Pago> findByMatriculaIdMatricula(Long idMatricula);
+public interface PagoRepository extends JpaRepository<Pago, Integer> {
+    List<Pago> findByMatriculaIdMatricula(Integer idMatricula);
     List<Pago> findByEstado(EstadoPago estado);
     List<Pago> findByTipo(TipoPago tipo);
 }
